requirements.txt
django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
firebase-admin>=6.2.0
python-dotenv>=1.0.0
requests>=2.31.0
google-cloud-firestore>=2.13.0
openai-whisper>=20230918
ffmpeg-python>=0.2.0
pydub>=0.25.1 